requirements.txt
appdirs==1.4.3
Django==1.11.7
djangorestframework==3.7.1
gunicorn==19.7.0
Markdown==2.6.9
olefile==0.44
packaging==16.8
Pillow==4.3.0
pkg-resources==0.0.0
psycopg2==2.7.3.2
pyparsing==2.2.0
reportlab==3.4.0
requests==2.18.4
six==1.11.0
